+++ src/main/java/com/codeit/mission/deokhugam/notification/repository/custom/NotificationRepositoryCustom.java
@@ -1,16 +1,16 @@
 package com.codeit.mission.deokhugam.notification.repository.custom;
 
-import com.codeit.mission.deokhugam.notification.dto.NotificationRequestQuery;
+import com.codeit.mission.deokhugam.notification.dto.request.NotificationRequestQuery;
 import com.codeit.mission.deokhugam.notification.entity.Notification;
 import java.util.UUID;
 import org.springframework.data.domain.Slice;
 
 public interface NotificationRepositoryCustom {
 
-    Slice<Notification> findByUserWithCursor(
-        UUID userId,
-        NotificationRequestQuery query
-    );
+  Slice<Notification> findByUserWithCursor(
+      UUID userId,
+      NotificationRequestQuery query
+  );
 
-    long countByUserId(UUID userId);
+  long countByUserId(UUID userId);
 }
