+++ src/main/java/com/codeit/mission/deokhugam/notification/repository/custom/NotificationRepositoryImpl.java
@@ -1,13 +1,13 @@
 package com.codeit.mission.deokhugam.notification.repository.custom;
 
 import com.codeit.mission.deokhugam.dashboard.DirectionEnum;
-import com.codeit.mission.deokhugam.notification.dto.NotificationRequestQuery;
+import com.codeit.mission.deokhugam.notification.dto.request.NotificationRequestQuery;
 import com.codeit.mission.deokhugam.notification.entity.Notification;
 import com.codeit.mission.deokhugam.notification.entity.QNotification;
 import com.querydsl.core.BooleanBuilder;
 import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import java.time.LocalDateTime;
+import java.time.Instant;
 import java.time.ZoneOffset;
 import java.util.List;
 import java.util.UUID;
@@ -19,74 +19,74 @@ import org.springframework.data.domain.SliceImpl;
 @RequiredArgsConstructor
 public class NotificationRepositoryImpl implements NotificationRepositoryCustom {
 
-    private final JPAQueryFactory queryFactory;
-    QNotification notification = QNotification.notification;
+  private final JPAQueryFactory queryFactory;
+  QNotification notification = QNotification.notification;
 
-    @Override
-    public Slice<Notification> findByUserWithCursor(UUID userId, NotificationRequestQuery query) {
+  @Override
+  public Slice<Notification> findByUserWithCursor(UUID userId, NotificationRequestQuery query) {
 
-        int limit = query.getLimitOrDefault();
+    int limit = query.getLimitOrDefault();
 
-        BooleanBuilder builder = new BooleanBuilder();
-        builder.and(notification.user.id.eq(userId));
+    BooleanBuilder builder = new BooleanBuilder();
+    builder.and(notification.user.id.eq(userId));
 
-        // after (createdAt 기준 커서)
-        if (query.after() != null) {
-            LocalDateTime afterTime = LocalDateTime.ofInstant(query.after(), ZoneOffset.UTC);
+    // after (createdAt 기준 커서)
+    if (query.after() != null) {
+      Instant afterTime = query.after();
 
-            if (isDesc(query)) {
-                builder.and(notification.createdAt.lt(afterTime));
-            } else {
-                builder.and(notification.createdAt.gt(afterTime));
-            }
-        }
-
-        // cursor
-        if (query.cursor() != null) {
-            LocalDateTime cursorTime = LocalDateTime.ofInstant(query.cursor(), ZoneOffset.UTC);
-
-            if (isDesc(query)) {
-                builder.and(notification.createdAt.lt(LocalDateTime.from(cursorTime)));
-            } else {
-                builder.and(notification.createdAt.gt(LocalDateTime.from(cursorTime)));
-            }
-        }
-
-        List<Notification> result = queryFactory
-            .selectFrom(notification)
-            .where(builder)
-            .orderBy(getOrderSpecifier(query))
-            .limit(limit + 1)
-            .fetch();
-
-        boolean hasNext = result.size() > limit;
-
-        if (hasNext) {
-            result = result.subList(0, limit);
-        }
-
-        return new SliceImpl<>(result, PageRequest.of(0, limit), hasNext);
+      if (isDesc(query)) {
+        builder.and(notification.createdAt.lt(afterTime));
+      } else {
+        builder.and(notification.createdAt.gt(afterTime));
+      }
     }
 
-    @Override
-    public long countByUserId(UUID userId) {
-        Long count = queryFactory
-            .select(notification.count())
-            .from(notification)
-            .where(notification.user.id.eq(userId))
-            .fetchOne();
+    // cursor
+    if (query.cursor() != null) {
+      Instant cursorTime = query.after();
 
-        return count == null ? 0L : count;
+      if (isDesc(query)) {
+        builder.and(notification.createdAt.lt(Instant.from(cursorTime)));
+      } else {
+        builder.and(notification.createdAt.gt(Instant.from(cursorTime)));
+      }
     }
 
-    private boolean isDesc(NotificationRequestQuery query) {
-        return query.direction() == null || query.direction().equals(DirectionEnum.DESC);
+    List<Notification> result = queryFactory
+        .selectFrom(notification)
+        .where(builder)
+        .orderBy(getOrderSpecifier(query))
+        .limit(limit + 1)
+        .fetch();
+
+    boolean hasNext = result.size() > limit;
+
+    if (hasNext) {
+      result = result.subList(0, limit);
     }
 
-    private OrderSpecifier<?> getOrderSpecifier(NotificationRequestQuery query) {
-        return isDesc(query)
-            ? notification.createdAt.desc()
-            : notification.createdAt.asc();
-    }
+    return new SliceImpl<>(result, PageRequest.of(0, limit), hasNext);
+  }
+
+  @Override
+  public long countByUserId(UUID userId) {
+    Long count = queryFactory
+        .select(notification.count())
+        .from(notification)
+        .where(notification.user.id.eq(userId))
+        .fetchOne();
+
+    return count == null ? 0L : count;
+  }
+
+  private boolean isDesc(NotificationRequestQuery query) {
+    return query.direction() == null || query.direction().equals(DirectionEnum.DESC);
+  }
+
+  private OrderSpecifier<?> getOrderSpecifier(NotificationRequestQuery query) {
+    return isDesc(query)
+        ? notification.createdAt.desc()
+        : notification.createdAt.asc();
+  }
 
 }
