+++ src/test/java/com/codeit/mission/deokhugam/notification/repository/NotificationRepositoryTest.java
@@ -6,15 +6,14 @@ import static org.assertj.core.api.Assertions.assertThat;
 import com.codeit.mission.deokhugam.book.entity.Book;
 import com.codeit.mission.deokhugam.config.QuerydslConfig;
 import com.codeit.mission.deokhugam.dashboard.DirectionEnum;
-import com.codeit.mission.deokhugam.notification.dto.NotificationRequestQuery;
+import com.codeit.mission.deokhugam.notification.dto.request.NotificationRequestQuery;
 import com.codeit.mission.deokhugam.notification.entity.Notification;
 import com.codeit.mission.deokhugam.review.entity.Review;
 import com.codeit.mission.deokhugam.user.entity.User;
 import jakarta.persistence.EntityManager;
 import java.time.Instant;
-import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.time.ZoneOffset;
+import java.time.ZoneId;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
@@ -44,31 +43,32 @@ public class NotificationRepositoryTest {
 
   @BeforeEach
   void setUp() {
-    LocalDateTime BASE_TIME =
-      LocalDateTime.of(2026, 1, 1, 0, 0, 0);
+    Instant BASE_TIME = LocalDateTime.of(2026, 1, 1, 0, 0, 0)
+        .atZone(ZoneId.of("Asia/Seoul"))
+        .toInstant();
 
     user = User.builder()
-      .email("test@test.com")
-      .nickname("test")
-      .password("password")
-      .build();
+        .email("test@test.com")
+        .nickname("test")
+        .password("password")
+        .build();
 
     book = Book.builder()
-      .rating(5)
-      .title("제목")
-      .publishedDate(LocalDate.now())
-      .author("작가")
-      .publisher("출판사")
-      .description("설명")
-      .isbn("")
-      .build();
+        .rating(5)
+        .title("제목")
+        .publishedDate(Instant.now())
+        .author("작가")
+        .publisher("출판사")
+        .description("설명")
+        .isbn("")
+        .build();
 
     review = Review.builder()
-      .rating(5)
-      .user(user)
-      .content("리뷰 내용")
-      .book(book)
-      .build();
+        .rating(5)
+        .user(user)
+        .content("리뷰 내용")
+        .book(book)
+        .build();
 
     em.persist(user);
     em.persist(book);
@@ -77,19 +77,19 @@ public class NotificationRepositoryTest {
     notificationList = new ArrayList<>();
     for (int i = 0; i < 30; i++) {
       Notification notification = Notification.builder()
-        .user(user)
-        .message(i + "번째 알림")
-        .review(review)
-        .reviewContent(review.getContent())
-        .message("알림 메시지 " + i)
-        .confirmed(i % 2 == 0)
-        .build();
+          .user(user)
+          .message(i + "번째 알림")
+          .review(review)
+          .reviewContent(review.getContent())
+          .message("알림 메시지 " + i)
+          .confirmed(i % 2 == 0)
+          .build();
 
       // reflection으로 createAt 강제 세팅
       ReflectionTestUtils.setField(
-        notification,
-        "createdAt",
-        BASE_TIME.minusSeconds(i)
+          notification,
+          "createdAt",
+          BASE_TIME.minusSeconds(i)
       );
 
       notificationList.add(notification);
@@ -107,12 +107,12 @@ public class NotificationRepositoryTest {
     UUID userId = user.getId();
 
     NotificationRequestQuery query = NotificationRequestQuery.builder()
-      .direction(DirectionEnum.DESC)
-      .build();
+        .direction(DirectionEnum.DESC)
+        .build();
 
     // when
     Slice<Notification> result =
-      notificationRepository.findByUserWithCursor(userId, query);
+        notificationRepository.findByUserWithCursor(userId, query);
 
     // then
     assertThat(result.getContent()).hasSize(20);
@@ -124,34 +124,33 @@ public class NotificationRepositoryTest {
   void findNextPage() {
     // given
     NotificationRequestQuery firstQuery = NotificationRequestQuery.builder()
-      .direction(DirectionEnum.DESC)
-      .limit(20)
-      .build();
+        .direction(DirectionEnum.DESC)
+        .limit(20)
+        .build();
 
     Slice<Notification> firstPage =
-      notificationRepository.findByUserWithCursor(user.getId(), firstQuery);
+        notificationRepository.findByUserWithCursor(user.getId(), firstQuery);
 
     Instant cursor = firstPage.getContent().get(19)
-      .getCreatedAt()
-      .toInstant(ZoneOffset.UTC);
+        .getCreatedAt();
 
     NotificationRequestQuery secondQuery = NotificationRequestQuery.builder()
-      .direction(DirectionEnum.DESC)
-      .after(cursor)
-      .limit(20)
-      .build();
+        .direction(DirectionEnum.DESC)
+        .after(cursor)
+        .limit(20)
+        .build();
 
     // when
     Slice<Notification> secondPage =
-      notificationRepository.findByUserWithCursor(user.getId(), secondQuery);
+        notificationRepository.findByUserWithCursor(user.getId(), secondQuery);
 
     // then
     assertThat(secondPage.getContent()).hasSize(10);
 
     secondPage.getContent().forEach(n -> {
-        Instant createdAt = n.getCreatedAt().toInstant(ZoneOffset.UTC);
-        assertThat(createdAt).isBefore(cursor);
-      }
+          Instant createdAt = n.getCreatedAt();
+          assertThat(createdAt).isBefore(cursor);
+        }
     );
   }
 
@@ -160,13 +159,13 @@ public class NotificationRepositoryTest {
   void findLastPage() {
     // given
     NotificationRequestQuery query = NotificationRequestQuery.builder()
-      .direction(DirectionEnum.DESC)
-      .limit(50)
-      .build();
+        .direction(DirectionEnum.DESC)
+        .limit(50)
+        .build();
 
     // when
     Slice<Notification> result =
-      notificationRepository.findByUserWithCursor(user.getId(), query);
+        notificationRepository.findByUserWithCursor(user.getId(), query);
 
     // then
     assertThat(result.hasNext()).isFalse();
@@ -177,19 +176,19 @@ public class NotificationRepositoryTest {
   void findWithAsc() {
     // given
     NotificationRequestQuery query = NotificationRequestQuery.builder()
-      .direction(DirectionEnum.ASC)
-      .build();
+        .direction(DirectionEnum.ASC)
+        .build();
 
     // when
     Slice<Notification> result =
-      notificationRepository.findByUserWithCursor(user.getId(), query);
+        notificationRepository.findByUserWithCursor(user.getId(), query);
 
     // then
     List<Notification> content = result.getContent();
 
     for (int i = 0; i < content.size() - 1; i++) {
       assertThat(content.get(i).getCreatedAt())
-        .isBeforeOrEqualTo(content.get(i + 1).getCreatedAt());
+          .isBeforeOrEqualTo(content.get(i + 1).getCreatedAt());
     }
   }
 
@@ -215,12 +214,12 @@ public class NotificationRepositoryTest {
 
     // 실제로 confirm이 true로 바뀌었는지 검증
     List<Notification> notifications = em
-      .createQuery("select n from Notification n where n.user.id = :userId", Notification.class)
-      .setParameter("userId", user.getId())
-      .getResultList();
+        .createQuery("select n from Notification n where n.user.id = :userId", Notification.class)
+        .setParameter("userId", user.getId())
+        .getResultList();
 
     assertThat(notifications)
-      .allMatch(Notification::isConfirmed);
+        .allMatch(Notification::isConfirmed);
 
   }
 }
